+++ src/TodoItem/index.js
@@ -1,5 +1,5 @@
-import { CompleteIcon } from './CompleteIcon'
-import { DeleteIcon } from './deleteIcon'
+import { CompleteIcon } from '../TodoIcon/CompleteIcon'
+import { DeleteIcon } from '../TodoIcon/deleteIcon'
 import './TodoItem.css';
 
 function TodoItem(props) {
